Making tests smaller
