+++ tests/integration/mci/test_invalid_dates_numbers.py
@@ -5,12 +5,9 @@ from tests.integration.mci import mci_test_urls
 
 class TestInvalidDateNumber(IntegrationTestCase):
 
-    def test_invalid_date_number_205(self):
-        self.invalid_date_number('0205', '1')
-
-    def invalid_date_number(self, form_type_id, eq_id):
+    def test_correct_invalid_date(self):
         # Get a token
-        token = create_token(form_type_id, eq_id)
+        token = create_token('0205', '1')
         resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
         self.assertEqual(resp.status_code, 200)
 
@@ -27,15 +24,6 @@ class TestInvalidDateNumber(IntegrationTestCase):
         resp = self.client.post(mci_test_urls.MCI_0205_INTRODUCTION, data=post_data, follow_redirects=False)
         self.assertEqual(resp.status_code, 302)
 
-        # We proceed to the questionnaire
-        resp = self.client.get(mci_test_urls.MCI_0205_BLOCK1, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-
-        # We are in the Questionnaire
-        content = resp.get_data(True)
-        self.assertRegex(content, "What are the dates of the sales period you are reporting for?")
-        self.assertRegex(content, ">Save and continue<")
-
         form_data = {
             # Start Date
             "period-from-day": "01",
@@ -51,11 +39,11 @@ class TestInvalidDateNumber(IntegrationTestCase):
             "action[save_continue]": "Save &amp; Continue"
         }
 
-        # We submit the form with an invalid date
+        # We submit the form with the dates the same
         resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
         self.assertEqual(resp.status_code, 200)
         content = resp.get_data(True)
-        self.assertRegex(content, "The date entered is not valid.  Please correct your answer.")
+        self.assertRegex(content, "The &#39;period to&#39; date must be different to the &#39;period from&#39; date.")
 
         form_data = {
             # Start Date
@@ -68,96 +56,6 @@ class TestInvalidDateNumber(IntegrationTestCase):
             "period-to-year": "2017",
             # Total Turnover
             "total-retail-turnover": "rubbish",
-            # User action
-            "action[save_continue]": "Save &amp; Continue"
-        }
-
-        # We submit the form without a valid turnover value
-        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-        content = resp.get_data(True)
-        self.assertRegex(content, "Please only enter whole numbers into the field.")
-
-        form_data = {
-            # Start Date
-            "period-from-day": "01",
-            "period-from-month": "01",
-            "period-from-year": "2016",
-            # End Date
-            "period-to-day": "01",
-            "period-to-month": "01",
-            "period-to-year": "",
-            # Total Turnover
-            "total-retail-turnover": "10000",
-            # User action
-            "action[save_continue]": "Save &amp; Continue"
-        }
-
-        # We submit the form without a valid 2nd date
-        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-        content = resp.get_data(True)
-        self.assertRegex(content, "The date entered is not valid")
-
-        # We try to access the submission page without correction
-        resp = self.client.get(mci_test_urls.MCI_0205_SUMMARY, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-        content = resp.get_data(True)
-        self.assertRegex(content, "What are the dates of the sales period you are reporting for?")
-
-        form_data = {
-            # Start Date
-            "period-from-day": "01",
-            "period-from-month": "01",
-            "period-from-year": "2017",
-            # End Date
-            "period-to-day": "01",
-            "period-to-month": "01",
-            "period-to-year": "2016",
-            # Total Turnover
-            "total-retail-turnover": "10000",
-            # User action
-            "action[save_continue]": "Save &amp; Continue"
-        }
-
-        # We submit the form with the front date later then the to date
-        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-        content = resp.get_data(True)
-        self.assertRegex(content, "The &#39;period to&#39; date cannot be before the &#39;period from&#39; date.")
-
-        form_data = {
-            # Start Date
-            "period-from-day": "01",
-            "period-from-month": "01",
-            "period-from-year": "2016",
-            # End Date
-            "period-to-day": "01",
-            "period-to-month": "01",
-            "period-to-year": "2016",
-            # Total Turnover
-            "total-retail-turnover": "100000",
-            # User action
-            "action[save_continue]": "Save &amp; Continue"
-        }
-
-        # We submit the form with the dates the same
-        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
-        self.assertEqual(resp.status_code, 200)
-        content = resp.get_data(True)
-        self.assertRegex(content, "The &#39;period to&#39; date must be different to the &#39;period from&#39; date.")
-
-        form_data = {
-            # Start Date
-            "period-from-day": "01",
-            "period-from-month": "4",
-            "period-from-year": "2016",
-            # End Date
-            "period-to-day": "30",
-            "period-to-month": "04",
-            "period-to-year": "2016",
-            # Total Turnover
-            "total-retail-turnover": "100000",
             # User action
             "action[save_continue]": "Save &amp; Continue"
         }
@@ -190,3 +88,130 @@ class TestInvalidDateNumber(IntegrationTestCase):
         # We are on the thank you page
         content = resp.get_data(True)
         self.assertRegex(content, '<title>Submission Successful</title>')
+
+    def test_invalid_same_date(self):
+        # Get a token
+        token = create_token('0205', '1')
+        resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        form_data = {
+            # Start Date
+            "period-from-day": "01",
+            "period-from-month": "01",
+            "period-from-year": "2016",
+            # End Date
+            "period-to-day": "01",
+            "period-to-month": "01",
+            "period-to-year": "",
+            # Total Turnover
+            "total-retail-turnover": "10000",
+            # User action
+            "action[save_continue]": "Save &amp; Continue"
+        }
+        # We submit the form with the dates the same
+        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        content = resp.get_data(True)
+        self.assertRegex(content, "The &#39;period to&#39; date must be different to the &#39;period from&#39; date.")
+
+    def test_invalid_date_range(self):
+        # Get a token
+        token = create_token('0205', '1')
+        resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+
+        form_data = {
+            # Start Date
+            "period-from-day": "01",
+            "period-from-month": "01",
+            "period-from-year": "2017",
+            # End Date
+            "period-to-day": "01",
+            "period-to-month": "01",
+            "period-to-year": "2016",
+            # Total Turnover
+            "total-retail-turnover": "10000",
+            # User action
+            "action[save_continue]": "Save &amp; Continue"
+        }
+        # We submit the form with the front date later then the to date
+        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        content = resp.get_data(True)
+        self.assertRegex(content, "The &#39;period to&#39; date cannot be before the &#39;period from&#39; date.")
+
+    def test_invalid_year(self):
+        # Get a token
+        token = create_token('0205', '1')
+        resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        form_data = {
+            # Start Date
+            "period-from-day": "01",
+            "period-from-month": "01",
+            "period-from-year": "2016",
+            # End Date
+            "period-to-day": "01",
+            "period-to-month": "01",
+            "period-to-year": "2016",
+            # Total Turnover
+            "total-retail-turnover": "100000",
+            # User action
+            "action[save_continue]": "Save &amp; Continue"
+        }
+        # We submit the form without a valid 2nd date
+        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        content = resp.get_data(True)
+        self.assertRegex(content, "The date entered is not valid")
+
+    def test_invalid_integer(self):
+        # Get a token
+        token = create_token('0205', '1')
+        resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        form_data = {
+            # Start Date
+            "period-from-day": "01",
+            "period-from-month": "4",
+            "period-from-year": "2016",
+            # End Date
+            "period-to-day": "30",
+            "period-to-month": "04",
+            "period-to-year": "2016",
+            # Total Turnover
+            "total-retail-turnover": "100000",
+            # User action
+            "action[save_continue]": "Save &amp; Continue"
+        }
+        # We submit the form without a valid turnover value
+        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        content = resp.get_data(True)
+        self.assertRegex(content, "Please only enter whole numbers into the field.")
+
+    def test_invalid_date_number(self):
+        # Get a token
+        token = create_token('0205', '1')
+        resp = self.client.get('/session?token=' + token.decode(), follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+
+        form_data = {
+            # Start Date
+            "period-from-day": "01",
+            "period-from-month": "1",
+            "period-from-year": "",
+            # End Date
+            "period-to-day": "01",
+            "period-to-month": "01",
+            "period-to-year": "",
+            # Total Turnover
+            "total-retail-turnover": "100000",
+            # User action
+            "action[save_continue]": "Save &amp; Continue"
+        }
+        # We submit the form with an invalid date
+        resp = self.client.post(mci_test_urls.MCI_0205_BLOCK1, data=form_data, follow_redirects=True)
+        self.assertEqual(resp.status_code, 200)
+        content = resp.get_data(True)
+        self.assertRegex(content, "The date entered is not valid.  Please correct your answer.")
