+++ tests/integration/star_wars/test_check_boxes.py
@@ -5,7 +5,7 @@ from tests.integration.star_wars.star_wars_tests import StarWarsTestCase
 class TestEmptyCheckBoxes(StarWarsTestCase):
 
     def test_check_boxes_mandatory_empty(self):
-        self.login_and_check_introduction_text()
+        self.login()
 
         first_page = self.start_questionnaire_and_navigate_routing()
