+++ tests/integration/star_wars/test_conditional_display.py
@@ -1,5 +1,3 @@
-from werkzeug.datastructures import MultiDict
-
 from tests.integration.star_wars import star_wars_test_urls, BLOCK_2_DEFAULT_ANSWERS, BLOCK_8_DEFAULT_ANSWERS
 from tests.integration.star_wars.star_wars_tests import StarWarsTestCase
 
@@ -8,32 +6,31 @@ class TestConditionalDisplay(StarWarsTestCase):
 
     def test_conditional_display_questions_present(self):
 
-        self.login_and_check_introduction_text()
+        self.login()
 
         first_page = self.start_questionnaire_and_navigate_routing()
 
-        form_data = MultiDict()
-        form_data.add("cccfe681-9969-4175-8ac3-98184ab58423", "Bothans")
-        form_data.add("6cf5c72a-c1bf-4d0c-af6c-d0f07bc5b65b", "234")
-        form_data.add("92e49d93-cbdc-4bcb-adb2-0e0af6c9a07c", "40")
-        form_data.add("pre49d93-cbdc-4bcb-adb2-0e0af6c9a07c", "1370")
+        form_data = {
+            "cccfe681-9969-4175-8ac3-98184ab58423": "Bothans",
+            "6cf5c72a-c1bf-4d0c-af6c-d0f07bc5b65b": "234",
+            "92e49d93-cbdc-4bcb-adb2-0e0af6c9a07c": "40",
+            "pre49d93-cbdc-4bcb-adb2-0e0af6c9a07c": "1370",
 
-        form_data.add("a5dc09e8-36f2-4bf4-97be-c9e6ca8cbe0d", "Elephant")
-        form_data.add("7587eb9b-f24e-4dc0-ac94-66118b896c10", "Luke, I am your father")
+            "a5dc09e8-36f2-4bf4-97be-c9e6ca8cbe0d": "Elephant",
+            "7587eb9b-f24e-4dc0-ac94-66118b896c10": "Luke, I am your father",
 
-        form_data.add("9587eb9b-f24e-4dc0-ac94-66117b896c10", 'Luke Skywalker')
-        form_data.add("9587eb9b-f24e-4dc0-ac94-66117b896c10", 'Yoda')
-        form_data.add("9587eb9b-f24e-4dc0-ac94-66117b896c10", 'Qui-Gon Jinn')
+            "9587eb9b-f24e-4dc0-ac94-66117b896c10": ["Luke Skywalker", "Yoda", "Qui-Gon Jinn"],
 
-        form_data.add("6fd644b0-798e-4a58-a393-a438b32fe637-day", "28")
-        form_data.add("6fd644b0-798e-4a58-a393-a438b32fe637-month", "05")
-        form_data.add("6fd644b0-798e-4a58-a393-a438b32fe637-year", "1983")
+            "6fd644b0-798e-4a58-a393-a438b32fe637-day": "28",
+            "6fd644b0-798e-4a58-a393-a438b32fe637-month": "05",
+            "6fd644b0-798e-4a58-a393-a438b32fe637-year": "1983",
 
-        form_data.add("06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-day", "29")
-        form_data.add("06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-month", "05")
-        form_data.add("06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-year", "1983")
+            "06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-day": "29",
+            "06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-month": "05",
+            "06a6a4b7-6ce4-4687-879d-3443cd8e2ff0-year": "1983",
 
-        form_data.add("action[save_continue]", "Save &amp; Continue")
+            "action[save_continue]": "Save &amp; Continue"
+        }
 
         # We submit the form
         resp = self.submit_page(first_page, form_data)
@@ -85,7 +82,7 @@ class TestConditionalDisplay(StarWarsTestCase):
 
     def test_conditional_display_questions_non_present(self):
 
-        self.login_and_check_introduction_text()
+        self.login()
 
         first_page = self.start_questionnaire_and_navigate_routing()
