+++ tests/integration/star_wars/test_confirmation_page.py
@@ -1,5 +1,4 @@
-from werkzeug.datastructures import MultiDict
-
+from tests.integration.create_token import create_token
 from tests.integration.star_wars import star_wars_test_urls
 from tests.integration.star_wars.star_wars_tests import StarWarsTestCase
 
@@ -74,3 +73,35 @@ class TestConfirmationPage(StarWarsTestCase):
         self.assertNotEqual(resp.location, confirmation_page)
 
         self.complete_survey(confirmation_page, 'rogue_one')
+
+    def rogue_one_login_and_check_introduction_text(self):
+        token = create_token('rogue_one', '0')
+        response = self.get_first_page(token)
+        self.rogue_one_check_introduction_text(response)
+
+    def rogue_one_check_introduction_text(self, response):
+        content = response.get_data(True)
+        self.assertIn('<title>Introduction</title>', content)
+        self.assertRegex(content, '(?s)Rogue One')
+
+    def rogue_one_check_character_page(self, page):
+        content = self.retrieve_content(page)
+        self.assertIn('Who do you want to know more about?', content)
+        self.assertIn('Jyn Erso', content)
+        self.assertIn('ca3ce3a3-ae44-4e30-8f85-5b6a7a2fb23c-3', content)
+
+    def rogue_one_check_description_page(self, page):
+        content = self.retrieve_content(page)
+        self.assertIn('An accomplished Rebel Alliance Intelligence Officer', content)
+        self.assertIn('Do you like this page?', content)
+        self.assertIn('3f1f1bb7-2452-4f8d-ac7a-735ea5d4517f-2', content)
+
+    def rogue_one_check_takings_page(self, page):
+        content = self.retrieve_content(page)
+        self.assertIn('In millions, how much do you think this film will take?', content)
+        self.assertIn('a04a516d-502d-4068-bbed-a43427c68cd9', content)
+
+    def rogue_one_check_confirmation_page(self, page):
+        content = self.retrieve_content(page)
+        self.assertIn('Summary', content)
+        self.assertIn('Please check carefully before submission', content)
